+++ composer.json
@@ -69,7 +69,7 @@
         "hyperf/grpc-server": "~3.0.0",
         "hyperf/json-rpc": "~3.0.0",
         "hyperf/guzzle": "~3.0.0",
-        "hyperf/http-server": "~3.0.20",
+        "hyperf/http-server": "~3.0.45",
         "hyperf/http2-client": "~3.0.0",
         "hyperf/ide-helper": "~3.0.0",
         "hyperf/kafka": "~3.0.0",
@@ -81,7 +81,7 @@
         "hyperf/redis": "~3.0.0",
         "hyperf/rpc": "~3.0.0",
         "hyperf/rpc-client": "~3.0.0",
-        "hyperf/rpc-server": "~3.0.0",
+        "hyperf/rpc-server": "~3.0.45",
         "hyperf/scout": "~3.0.0",
         "hyperf/session": "~3.0.0",
         "hyperf/stringable": "~3.0.40",
