+++ src/sentry/publish/sentry.php
@@ -95,6 +95,7 @@ return [
             'guzzle' => env('SENTRY_TRACING_ENABLE_GUZZLE', true),
             'rpc' => env('SENTRY_TRACING_ENABLE_RPC', true),
             'redis' => env('SENTRY_TRACING_ENABLE_REDIS', true),
+            'request' => env('SENTRY_TRACING_ENABLE_REQUEST', true),
             'sql_queries' => env('SENTRY_TRACING_ENABLE_SQL_QUERIES', true),
             'missing_routes' => env('SENTRY_TRACING_ENABLE_MISSING_ROUTES', true),
         ],
