+++ src/sentry/src/ConfigProvider.php
@@ -53,6 +53,7 @@ class ConfigProvider
                 Listener\RequestExceptionListener::class,
                 Listener\SetRequestLifecycleListener::class,
                 Tracing\Listener\TracingDbQueryListener::class,
+                Tracing\Listener\TracingRequestListener::class,
             ],
             'annotations' => [
                 'scan' => [
