+++ src/sentry/src/Listener/SetRequestLifecycleListener.php
@@ -15,7 +15,8 @@ use FriendsOfHyperf\Sentry\Switcher;
 use Hyperf\Contract\ConfigInterface;
 use Hyperf\Event\Contract\ListenerInterface;
 use Hyperf\Framework\Event\BootApplication;
-use Hyperf\HttpServer\Server;
+use Hyperf\HttpServer\Server as HttpServer;
+use Hyperf\RpcServer\Server as RpcServer;
 use Hyperf\Server\Event;
 
 class SetRequestLifecycleListener implements ListenerInterface
@@ -35,7 +36,14 @@ class SetRequestLifecycleListener implements ListenerInterface
 
     public function process(object $event): void
     {
-        if (! $this->switcher->isEnable('request')) {
+        if (! $this->config->has($key = 'sentry.tracing.enable.request')) {
+            $this->config->set($key, true);
+        }
+
+        if (
+            ! $this->switcher->isEnable('request')
+            && ! $this->switcher->isTracingEnable('request')
+        ) {
             return;
         }
 
@@ -43,13 +51,16 @@ class SetRequestLifecycleListener implements ListenerInterface
 
         foreach ($servers as &$server) {
             $callbacks = $server['callbacks'] ?? [];
-            $onRequestHandler = $callbacks[Event::ON_REQUEST][0] ?? null;
+            $handler = $callbacks[Event::ON_REQUEST][0] ?? $callbacks[Event::ON_RECEIVE][0] ?? null;
 
-            if (! $onRequestHandler) {
+            if (! $handler) {
                 continue;
             }
 
-            if ($onRequestHandler == Server::class || is_a($onRequestHandler, Server::class)) {
+            if (
+                is_a($handler, HttpServer::class, true)
+                || is_a($handler, RpcServer::class, true)
+            ) {
                 $server['options'] ??= [];
                 $server['options']['enable_request_lifecycle'] = true;
             }
